[#16] refactor : 클래스 분리, 변수명 수정, 중복코드 제거
- 하나의 디렉토리를 여러 디렉토리로 세분화

- 하나의 클래스에서 다양한 처리를 하는것을 세분화

- 변수명을 조금 더 명확하게 수정

- 중복코드 수정 (현재, URL 만드는 부분까지는 별도의 클래스에서 처리하도록 구현)

#16

# 꼬릿말은 아래에 작성 : ex) #이슈 번호

# --COMMIT END--
# <타입> 리스트
# feat : 기능 (새로운 기능)
# fix : 버그 (버그 수정)
# refactor : 리팩토링
# style : 스타일(코드 형식, 세미콜론 추가 : 비지니스 로직 변경 X)
# docs : 문서 (문서 추가, 수정, 삭제)
# test : 테스트 (테스트 코드 추가, 수정, 삭제 : 비지니스 로직 변경 X)
# chroe : 기타 변경사항(빌드 스크립트 수정 등)
-------------------------
#	제목 첫 글자 대문자
#	제목은 명령문
#	제목 끝에 마침표(.) 금지
#	제목과 본문을 한 줄 띄워 분리하기
#	본문은 "어떻게" 보다 "무엇을" , "왜"를 설명한다
#	본문에 여러줄의 메시지를 작성할 땐 "-"로 구분
-------------------------
