+++ BE/src/main/java/com/codesquad/dust7/domain/StationInfoData.java
@@ -1,16 +1,16 @@
-package com.codesquad.dust7;
+package com.codesquad.dust7.domain;
 
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
 
 @Getter @Setter @ToString
-public class StationInformation {
+public class StationInfoData {
     private String stationName;
     private String address;
     private String length;
 
-    public StationInformation(String stationName, String address, String length) {
+    public StationInfoData(String stationName, String address, String length) {
         this.stationName = stationName;
         this.address = address;
         this.length = length;
